Chore: moved UserController package
