+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.zerobase.hseungho.restaurantreservation.service.controller.user;
+package com.zerobase.hseungho.restaurantreservation.service.controller;
 
 import com.zerobase.hseungho.restaurantreservation.service.appservice.UserService;
 import com.zerobase.hseungho.restaurantreservation.service.dto.external.user.CheckUsingResourceAvailable;
