+++ src/test/java/com/zerobase/hseungho/restaurantreservation/service/controller/user/UserControllerTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zerobase.hseungho.restaurantreservation.global.config.SecurityConfiguration;
 import com.zerobase.hseungho.restaurantreservation.global.security.jwt.JwtAuthenticationFilter;
 import com.zerobase.hseungho.restaurantreservation.global.util.SeoulDateTime;
+import com.zerobase.hseungho.restaurantreservation.service.controller.UserController;
 import com.zerobase.hseungho.restaurantreservation.service.appservice.UserServiceImpl;
 import com.zerobase.hseungho.restaurantreservation.service.dto.external.user.SignUp;
 import com.zerobase.hseungho.restaurantreservation.service.dto.internal.user.UserDto;
